+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/entity/SupportSlime.kt
@@ -6,8 +6,8 @@ import dev.aaronhowser.mods.geneticsresequenced.config.ServerConfig
 import dev.aaronhowser.mods.geneticsresequenced.datagen.ModLanguageProvider
 import dev.aaronhowser.mods.geneticsresequenced.datagen.ModLanguageProvider.Companion.toComponent
 import dev.aaronhowser.mods.geneticsresequenced.entity.goals.SupportSlimeAttackGoal
-import dev.aaronhowser.mods.geneticsresequenced.gene.ModGenes
 import dev.aaronhowser.mods.geneticsresequenced.registry.ModEntityTypes
+import dev.aaronhowser.mods.geneticsresequenced.registry.ModGenes
 import dev.aaronhowser.mods.geneticsresequenced.registry.ModItems
 import dev.aaronhowser.mods.geneticsresequenced.util.ModScheduler
 import dev.aaronhowser.mods.geneticsresequenced.util.OtherUtil.getUuidOrNull
@@ -64,10 +64,10 @@ class SupportSlime(
 
             if (!item.`is`(ModItems.FRIENDLY_SLIME_SPAWN_EGG.get())) return
 
-            if (!player.hasGene(ModGenes.slimyDeath)) {
+            if (!player.hasGene(ModGenes.SLIMY_DEATH.get())) {
                 player.sendSystemMessage(
                     ModLanguageProvider.Messages.SUPPORT_SLIME_CREATIVE.toComponent(
-                        ModGenes.slimyDeath.nameComponent
+                        ModGenes.SLIMY_DEATH.get().nameComponent
                     )
                 )
             }
@@ -104,7 +104,7 @@ class SupportSlime(
             boundingBox.inflate(10.0)
         ).sortedByDescending { distanceToSqr(it) }
 
-        val owner = nearbyLivingEntities.firstOrNull { it.hasGene(ModGenes.slimyDeath) }
+        val owner = nearbyLivingEntities.firstOrNull { it.hasGene(ModGenes.SLIMY_DEATH.get()) }
         if (owner != null) {
             setOwner(owner.uuid)
         } else {