+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/event/ModBusEvents.kt
@@ -2,9 +2,9 @@ package dev.aaronhowser.mods.geneticsresequenced.event
 
 import dev.aaronhowser.mods.geneticsresequenced.GeneticsResequenced
 import dev.aaronhowser.mods.geneticsresequenced.api.genes.Gene
+import dev.aaronhowser.mods.geneticsresequenced.api.genes.GeneRegistry
 import dev.aaronhowser.mods.geneticsresequenced.block.base.InventoryEnergyBlockEntity
 import dev.aaronhowser.mods.geneticsresequenced.entity.SupportSlime
-import dev.aaronhowser.mods.geneticsresequenced.gene.ModGenes
 import dev.aaronhowser.mods.geneticsresequenced.packet.ModPacketHandler
 import dev.aaronhowser.mods.geneticsresequenced.registry.ModAttributes
 import dev.aaronhowser.mods.geneticsresequenced.registry.ModBlockEntities
@@ -17,7 +17,6 @@ import net.neoforged.bus.api.SubscribeEvent
 import net.neoforged.fml.common.EventBusSubscriber
 import net.neoforged.fml.config.ModConfig
 import net.neoforged.fml.event.config.ModConfigEvent
-import net.neoforged.fml.event.lifecycle.FMLCommonSetupEvent
 import net.neoforged.neoforge.capabilities.Capabilities
 import net.neoforged.neoforge.capabilities.RegisterCapabilitiesEvent
 import net.neoforged.neoforge.energy.EnergyStorage
@@ -25,7 +24,7 @@ import net.neoforged.neoforge.event.entity.EntityAttributeCreationEvent
 import net.neoforged.neoforge.event.entity.EntityAttributeModificationEvent
 import net.neoforged.neoforge.items.IItemHandler
 import net.neoforged.neoforge.network.event.RegisterPayloadHandlersEvent
-import net.neoforged.neoforge.registries.DataPackRegistryEvent
+import net.neoforged.neoforge.registries.NewRegistryEvent
 
 @EventBusSubscriber(
     modid = GeneticsResequenced.ID,
@@ -34,10 +33,8 @@ import net.neoforged.neoforge.registries.DataPackRegistryEvent
 object ModBusEvents {
 
     @SubscribeEvent
-    fun registerDataPackRegistries(event: DataPackRegistryEvent.NewRegistry) {
-//        event.dataPackRegistry(Registries.ENCHANTMENT, Enchantment.DIRECT_CODEC)
-//        event.dataPackRegistry(Registries.RECIPE, Recipe.CODEC)
-//        event.dataPackRegistry(Registries.ADVANCEMENT, Advancement.CODEC)
+    fun onNewRegistry(event: NewRegistryEvent) {
+        event.register(GeneRegistry.GENE_REGISTRY)
     }
 
     @SubscribeEvent
@@ -48,11 +45,6 @@ object ModBusEvents {
         if (config.modId == GeneticsResequenced.ID && config.type == ModConfig.Type.SERVER) {
             Gene.checkDeactivationConfig()
         }
-    }
-
-    @SubscribeEvent
-    fun onCommonSetup(event: FMLCommonSetupEvent) {
-        ModGenes.registerDefaultGenes()
     }
 
     @SubscribeEvent