+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/recipe/brewing/GmoRecipe.kt
@@ -1,9 +1,9 @@
 package dev.aaronhowser.mods.geneticsresequenced.recipe.brewing
 
 import dev.aaronhowser.mods.geneticsresequenced.api.genes.Gene
-import dev.aaronhowser.mods.geneticsresequenced.gene.ModGenes
 import dev.aaronhowser.mods.geneticsresequenced.item.EntityDnaItem
 import dev.aaronhowser.mods.geneticsresequenced.item.GmoCell
+import dev.aaronhowser.mods.geneticsresequenced.registry.ModGenes
 import dev.aaronhowser.mods.geneticsresequenced.registry.ModItems
 import dev.aaronhowser.mods.geneticsresequenced.registry.ModPotions
 import dev.aaronhowser.mods.geneticsresequenced.util.OtherUtil
@@ -61,7 +61,7 @@ class GmoRecipe(
         GmoCell.setDetails(
             output,
             entityType,
-            ModGenes.basic
+            ModGenes.BASIC.get()
         )
 
         return output